Moving edit test case into edit namespace
