+++ osu.Game.Rulesets.Karaoke.Tests/Edit/TestSceneLyricLineStyle.cs
@@ -22,7 +22,7 @@ using osu.Game.Tests.Visual;
 using osuTK;
 using osuTK.Graphics;
 
-namespace osu.Game.Rulesets.Karaoke.Tests
+namespace osu.Game.Rulesets.Karaoke.Tests.Edit
 {
     [TestFixture]
     public class TestSceneLyricLineStyle : OsuTestScene
