+++ osu.Game.Rulesets.Karaoke.Tests/Edit/TestSceneLyricRubyRomaji.cs
@@ -24,7 +24,7 @@ using osu.Game.Rulesets.Karaoke.Tests.Beatmaps;
 using osu.Game.Rulesets.Karaoke.UI.Components;
 using osu.Game.Tests.Visual;
 
-namespace osu.Game.Rulesets.Karaoke.Tests
+namespace osu.Game.Rulesets.Karaoke.Tests.Edit
 {
     [TestFixture]
     public class TestSceneLyricRubyRomaji : OsuTestScene
