+++ osu.Game.Rulesets.Karaoke.Tests/Edit/TestSceneTranslate.cs
@@ -21,7 +21,7 @@ using osu.Game.Rulesets.Karaoke.UI.Components;
 using osu.Game.Tests.Visual;
 using osuTK;
 
-namespace osu.Game.Rulesets.Karaoke.Tests
+namespace osu.Game.Rulesets.Karaoke.Tests.Edit
 {
     [TestFixture]
     public class TestSceneTranslate : OsuTestScene
